# docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /v2/events/gallery:
    get:
      operationId: EventsController_getGallery
      summary: Get highlighted events with images for homepage gallery (public)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
  /v2/events:
    get:
      operationId: EventsController_getEvents
      summary: "Get events (filter: highlight, fromDate, toDate, category)"
      parameters:
        - 
          name: highlight
          required: false
          in: query
          schema:
            type: boolean
        - 
          name: fromDate
          required: false
          in: query
          schema:
            type: string
        - 
          name: toDate
          required: false
          in: query
          schema:
            type: string
        - 
          name: category
          required: false
          in: query
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    post:
      operationId: EventsController_create
      summary: Create event + emit notification queue
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateEventDto"
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
  /v2/events/{id}:
    get:
      operationId: EventsController_getById
      summary: Get event by ID
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    put:
      operationId: EventsController_update
      summary: Update event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDto"
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    delete:
      operationId: EventsController_delete
      summary: Delete event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
  /v2/events/{id}/attendees:
    get:
      operationId: EventsController_getAttendees
      summary: List attendees of an event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
    post:
      operationId: EventsController_addAttendee
      summary: Add/update an attendee (RSVP) for an event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAttendeeDto"
      responses:
        201:
          description: ""
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
  /v2/events/{id}/attendees/{memberId}:
    delete:
      operationId: EventsController_removeAttendee
      summary: Remove an attendee from an event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Removed
      tags:
        - Events (Sự kiện dòng họ)
      security:
        - 
          bearer: []
...
  /v2/graves:
    get:
      operationId: GravesController_getAllGraves
      summary: List all graves (filter by name)
      parameters:
        - 
          name: name
          required: false
          in: query
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraveResponseDto"
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
    post:
      operationId: GravesController_createGrave
      summary: Create new grave with GPS coordinates
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGraveDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraveResponseDto"
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
  /v2/graves/nearby:
    get:
      operationId: GravesController_getNearbyGraves
      summary: Find graves near coordinates (lat, lng, radiusKm)
      parameters:
        - 
          name: lat
          required: true
          in: query
          schema:
            type: number
        - 
          name: lng
          required: true
          in: query
          schema:
            type: number
        - 
          name: radiusKm
          required: false
          in: query
          schema:
            type: number
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraveResponseDto"
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
  /v2/graves/{id}:
    get:
      operationId: GravesController_getGraveById
      summary: Get grave by ID
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraveResponseDto"
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
    put:
      operationId: GravesController_updateGrave
      summary: Update grave info + coordinates
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGraveDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraveResponseDto"
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
    delete:
      operationId: GravesController_deleteGrave
      summary: Delete grave
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Graves (Mộ phần)
      security:
        - 
          bearer: []
  /v2/members/{memberId}/life-events:
    get:
      operationId: LifeEventsController_getByMember
      summary: Get a member life-event timeline (public)
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LifeEventResponseDto"
      tags:
        - Life Events (Quá trình sinh sống)
      security:
        - 
          bearer: []
    post:
      operationId: LifeEventsController_create
      summary: Add a life event to a member
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLifeEventDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifeEventResponseDto"
      tags:
        - Life Events (Quá trình sinh sống)
      security:
        - 
          bearer: []
  /v2/members/{memberId}/life-events/{id}:
    delete:
      operationId: LifeEventsController_delete
      summary: Delete a life event
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Life Events (Quá trình sinh sống)
      security:
        - 
          bearer: []
...
info:
  title: Family Tree API v2
  description: "## Vietnamese Family Tree Management API\n\n**Base URL:** `http://localhost:3002/v2`\n\n**Authentication:** Bearer JWT token (same as backend v1)\n\n### Modules\n- **Members** — CRUD thành viên + profile + avatar (async upload)\n- **Relationships** — Quan hệ mới (BIOLOGICAL/ADOPTED/SPOUSE) + tìm tổ tiên/con cháu\n- **Tree** — Cây gia phả full (Redis cache 1h) + subtree 4 thế hệ\n- **Anniversaries** — Ngày giỗ (filter by member, month, upcoming)\n- **Events** — Sự kiện dòng họ + notification queue\n- **Media** — Thư viện media: upload mọi loại (ảnh nén lossless bằng sharp; video/audio/tài liệu upload thẳng) → Vercel Blob; phân trang/lọc/tìm kiếm + thống kê + album\n- **Graves** — Mộ phần với tọa độ GPS + tìm kiếm gần nhất\n\n### Queue Jobs (QStash + Redis)\n| Queue | Trigger | Action |\n|-------|---------|--------|\n| avatar-upload | Create/Update member với file | Upload → Vercel Blob → cập nhật avatar_url |\n| media-process | Upload media | Ảnh: sharp nén lossless; khác: upload thẳng → Vercel Blob (off-request) |\n| report-generate | Create/Delete member | Tính stats → lưu Redis |\n| notification | New member/relationship/event | Log (Phase 1) |\n"
  version: 2.0.0
  contact:
    name: Family Tree Team
    url: ""
    email: ""
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
      description: Dán CHỈ token thô (lấy từ POST /v2/auth/login), KHÔNG kèm tiền tố "Bearer ".
  schemas:
...
    MemberTreeBriefDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: object
          nullable: true
          example: Chi nhánh Phú Thọ
      required:
        - id
...
    ProfileResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        fullName:
          type: string
          example: Nguyễn Văn A
        generation:
          type: number
          nullable: true
          example: 5
          description: Thế hệ NHẬP TAY — ưu tiên hơn giá trị suy ra. Xem `member.generation` cho giá trị hiệu lực.
        biography:
          type: object
          nullable: true
        occupation:
          type: object
          nullable: true
          example: Kỹ sư phần mềm
        address:
          type: object
          nullable: true
          example: Hà Nội
        notes:
          type: object
          nullable: true
        phone:
          type: object
          nullable: true
          example: 0988 123 456
        contactEmail:
          type: object
          nullable: true
          example: lienhe@gmail.com
        familyPosition:
          type: object
          nullable: true
          example: Con trưởng
          description: Vị trí trong gia đình
        roleTags:
          default: []
          description: Vai trò trong dòng họ
          type: array
          items:
            type: string
        committeeRole:
          type: object
          nullable: true
          description: Vai trò trong ban/hội đồng (clanRole)
        isCommittee:
          type: boolean
          default: false
        isNotable:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - member_id
        - fullName
        - roleTags
        - isCommittee
        - isNotable
        - created_at
        - updated_at
    MemberResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        private_id:
          type: object
          nullable: true
        name:
          type: string
          example: Nguyễn Văn A
        normalized_name:
          type: object
          nullable: true
          description: Tên đã bỏ dấu để tìm kiếm
        avatar_url:
          type: object
          nullable: true
          format: uri
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: object
          nullable: true
          example: 1990-01-01
        deathDate:
          type: object
          nullable: true
          example: 2020-12-31
        tree_id:
          type: object
          nullable: true
          format: uuid
          description: Chi nhánh — FK tới Tree
        generation:
          type: number
          nullable: true
          example: 5
          description: "Thế hệ (giá trị HIỆU LỰC, đã tính sẵn) = `profile.generation` nếu nhập tay, ngược lại suy ra từ quan hệ cha-con/vợ-chồng. Cập nhật bởi job nền `generation-recompute`. Muốn biết là nhập tay hay suy ra: `profile.generation != null`."
        generation_updated_at:
          type: string
          nullable: true
          format: date-time
          description: Lần cuối job nền ghi lại `generation`
        created_at:
          type: string
          format: date-time
        profile:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/ProfileResponseDto"
        tree:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberTreeBriefDto"
      required:
        - id
        - name
        - created_at
...
    EventResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Giỗ tổ dòng họ
        description:
          type: object
          nullable: true
        images:
          description: Ảnh sự kiện
          type: array
          items:
            type: string
            format: uri
        highlight:
          type: boolean
          default: false
          description: Nổi bật trên trang chủ
        date:
          type: string
          nullable: true
          format: date-time
        end_date:
          type: string
          nullable: true
          format: date-time
          description: Ngày kết thúc
        location:
          type: object
          nullable: true
          example: Nhà thờ họ
        category:
          type: object
          nullable: true
          example: Lễ
          description: Loại sự kiện
        isLunar:
          type: boolean
          default: false
          description: Ngày theo âm lịch
        attendeeCount:
          type: number
          example: 0
          description: Số người tham dự
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - images
        - highlight
        - isLunar
        - attendeeCount
        - created_at
        - updated_at
    CreateEventDto:
      type: object
      properties:
        title:
          type: string
          example: Họp Mặt Dòng Họ 2024
        description:
          type: string
        date:
          format: date-time
          type: string
          example: 2024-08-15T00:00:00.000Z
        end_date:
          format: date-time
          type: string
          example: 2024-08-16T00:00:00.000Z
          description: Ngày kết thúc (sự kiện nhiều ngày)
        location:
          type: string
        category:
          type: string
          example: Lễ
          description: Loại sự kiện
        isLunar:
          type: boolean
          default: false
          description: Ngày theo âm lịch
        highlight:
          type: boolean
          default: false
        images:
          type: array
          items:
            type: string
      required:
        - title
    UpdateEventDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          format: date-time
          type: string
        end_date:
          format: date-time
          type: string
        location:
          type: string
        category:
          type: string
        isLunar:
          type: boolean
        highlight:
          type: boolean
        images:
          type: array
          items:
            type: string
        existingImages:
          description: Ảnh hiện có được giữ lại (URL); ảnh mới upload sẽ nối vào sau
          type: array
          items:
            type: string
    AddAttendeeDto:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        rsvp_status:
          type: string
          example: going
          description: going | maybe | declined
      required:
        - member_id
...
    GraveResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Nghĩa trang dòng họ Nguyễn
        latitude:
          type: object
          example: 21.028511
          nullable: true
          description: Vĩ độ (null = chưa xác định GPS)
        longitude:
          type: object
          example: 105.804817
          nullable: true
          description: Kinh độ (null = chưa xác định GPS)
        description:
          type: object
          nullable: true
        member_id:
          type: object
          nullable: true
          format: uuid
          description: Thành viên liên kết
        deathDate:
          type: object
          nullable: true
          example: 1950-03-20
          description: Ngày mất
        relocationDate:
          type: object
          nullable: true
          example: 1960-11-05
          description: Ngày cải táng
        constructionDate:
          type: object
          nullable: true
          example: 1960-12-01
          description: Ngày xây dựng
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member:
          nullable: true
          description: Suy ra generation/gender/branch/clanRole từ đây
          allOf:
            - 
              $ref: "#/components/schemas/MemberResponseDto"
      required:
        - id
        - name
        - created_at
        - updated_at
    CreateGraveDto:
      type: object
      properties:
        name:
          type: string
          example: Mộ Nguyễn Văn A
        latitude:
          type: number
          example: 10.7769
          nullable: true
          description: Null = chưa xác định GPS
        longitude:
          type: number
          example: 106.7009
          nullable: true
          description: Null = chưa xác định GPS
        description:
          type: string
          example: Nghĩa trang Bình Hưng Hòa
        member_id:
          type: string
          format: uuid
          description: Liên kết mộ ↔ thành viên (suy ra generation/gender/branch...)
        deathDate:
          type: string
          example: 1950-03-20
          description: Ngày mất
        relocationDate:
          type: string
          example: 1960-11-05
          description: Ngày cải táng
        constructionDate:
          type: string
          example: 1960-12-01
          description: Ngày xây dựng
      required:
        - name
    UpdateGraveDto:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
        member_id:
          type: string
          format: uuid
        deathDate:
          type: string
        relocationDate:
          type: string
        constructionDate:
          type: string
    LifeEventResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        date:
          type: string
          example: 1990-05-12
        title:
          type: string
          example: Tốt nghiệp đại học
        description:
          type: object
          nullable: true
        category:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - member_id
        - date
        - title
        - created_at
    CreateLifeEventDto:
      type: object
      properties:
        date:
          type: string
          example: 1990-05-12
          description: Ngày diễn ra (chuỗi, giống Member.birthDate)
        title:
          type: string
          example: Tốt nghiệp đại học
        description:
          type: string
        category:
          type: string
          example: education
          description: Loại mốc (học vấn, sự nghiệp...)
      required:
        - date
        - title
...